:white_check_mark: [Test] 테스트 코드 리팩토링
